refactor: change each additional feature to "activate" pattern
adjusted project structure in preparation for next feature addition
