+++ src/commands/phpunit.ts
@@ -2,7 +2,7 @@ import { commands, ExtensionContext, Terminal, Uri, window, workspace } from 'co
 
 import path from 'path';
 import fs from 'fs';
-import { getMethods, getTestName } from '../parser/unittest';
+import { getMethods, getTestName } from '../parsers/unittest';
 
 let terminal: Terminal | undefined;
